+++ starters/nextjs-starter-approuter-ts/components/smart-components/client-components.ts
@@ -1,4 +1,5 @@
 import { SmartComponentMap } from "@pantheon-systems/cpub-react-sdk/components";
+import React from "react";
 import LeadCapture from "./lead-capture";
 import MediaPreview from "./media-preview";
 import { serverSmartComponentMap } from "./server-components";
@@ -7,15 +8,15 @@ import TileNavigation from "./tile-navigation";
 const clientSmartComponentMap: SmartComponentMap = {
   LEAD_CAPTURE: {
     ...serverSmartComponentMap.LEAD_CAPTURE,
-    reactComponent: LeadCapture,
+    reactComponent: LeadCapture as (props: unknown) => React.JSX.Element,
   },
   MEDIA_PREVIEW: {
     ...serverSmartComponentMap.MEDIA_PREVIEW,
-    reactComponent: MediaPreview,
+    reactComponent: MediaPreview as (props: unknown) => React.JSX.Element,
   },
   TILE_NAVIGATION: {
     ...serverSmartComponentMap.TILE_NAVIGATION,
-    reactComponent: TileNavigation,
+    reactComponent: TileNavigation as (props: unknown) => React.JSX.Element,
   },
 };