+++ starters/nextjs-starter-ts/components/grid.tsx
@@ -156,7 +156,6 @@ function GridItemCoverImage({
   return (
     <>
       {imageSrc != null ? (
-        // eslint-disable-next-line @next/next/no-img-element
         <img
           src={imageSrc}
           alt={imageAltText || undefined}