+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pantheon-systems/pcc-sdk-core
 
+## 5.2.0
+
+### Patch Changes
+
+- c30a021: Updating dependencies.
+
+## 5.2.0-beta.1
+
+### Patch Changes
+
+- c30a021: Updating dependencies.
+
 ## 5.2.0-beta.0
 
 ## 5.1.1